+++ log.go
@@ -1,21 +1,9 @@
 package log
 
-/*
-LevelLogger attaches a level to the log.
-
-Supported method signatures:
-
-	logger.Level("Message")
-	logger.Level("Message", map[string]interface{}{"key": "value"})
-	logger.Level(map[string]interface{}{"key": "value"})
-*/
-type LevelLogger interface {
-	// Debug logs on Debug level.
-	Debug(args ...interface{})
-
-	// Info logs on Info level.
-	Info(args ...interface{})
-
-	// Error logs on Error level.
-	Error(args ...interface{})
+// Logger is a simple logger interface based on go-kit.
+//
+// The same interface has been proposed to add to the standard library.
+// See https://docs.google.com/document/d/1shW9DZJXOeGbG9Mr9Us9MiaPqmlcVatD_D8lrOXRNMU/edit#
+type Logger interface {
+	Log(keyvals ...interface{}) error
 }