+++ logrus/logger_test.go
@@ -8,74 +8,13 @@ import (
 	"github.com/sirupsen/logrus/hooks/test"
 )
 
-func TestLogger_Debug_Empty(t *testing.T) {
-	testEmpty("Debug", t)
-}
-
-func TestLogger_Info_Empty(t *testing.T) {
-	testEmpty("Info", t)
-}
-
-func TestLogger_Error_Empty(t *testing.T) {
-	testEmpty("Error", t)
-}
-
-func TestLogger_Debug_Message(t *testing.T) {
-	testMessage("Debug", t)
-}
-
-func TestLogger_Info_Message(t *testing.T) {
-	testMessage("Info", t)
-}
-
-func TestLogger_Error_Message(t *testing.T) {
-	testMessage("Error", t)
-}
-
-func TestLogger_Debug_Context(t *testing.T) {
-	testContext("Debug", t)
-}
-
-func TestLogger_Info_Context(t *testing.T) {
-	testContext("Info", t)
-}
-
-func TestLogger_Error_Context(t *testing.T) {
-	testContext("Error", t)
-}
-
-func TestLogger_Debug_MessageAndContext(t *testing.T) {
-	testMessageAndContext("Debug", t)
-}
-
-func TestLogger_Info_MessageAndContext(t *testing.T) {
-	testMessageAndContext("Info", t)
-}
-
-func TestLogger_Error_MessageAndContext(t *testing.T) {
-	testMessageAndContext("Error", t)
-}
-
-func TestLogger_Debug_Invalid(t *testing.T) {
-	testInvalid("Debug", t)
-}
-
-func TestLogger_Info_Invalid(t *testing.T) {
-	testInvalid("Info", t)
-}
-
-func TestLogger_Error_Invalid(t *testing.T) {
-	testInvalid("Error", t)
-}
-
-func testEmpty(level string, t *testing.T) {
+func TestLogger_Empty(t *testing.T) {
 	testLogger, hook := test.NewNullLogger()
-	testLogger.Level = _logrus.DebugLevel
 	logger := &logrus.Logger{
 		Logger: testLogger,
 	}
 
-	invoke(logger, level, t)
+	logger.Log()
 
 	lastEntry := hook.LastEntry()
 
@@ -88,14 +27,13 @@ func testEmpty(level string, t *testing.T) {
 	}
 }
 
-func testMessage(level string, t *testing.T, args ...interface{}) {
+func TestLogger_Message(t *testing.T) {
 	testLogger, hook := test.NewNullLogger()
-	testLogger.Level = _logrus.DebugLevel
 	logger := &logrus.Logger{
 		Logger: testLogger,
 	}
 
-	invoke(logger, level, t, "Test message")
+	logger.Log("msg", "Test message")
 
 	lastEntry := hook.LastEntry()
 
@@ -108,14 +46,13 @@ func testMessage(level string, t *testing.T, args ...interface{}) {
 	}
 }
 
-func testContext(level string, t *testing.T, args ...interface{}) {
+func TestLogger_Context(t *testing.T) {
 	testLogger, hook := test.NewNullLogger()
-	testLogger.Level = _logrus.DebugLevel
 	logger := &logrus.Logger{
 		Logger: testLogger,
 	}
 
-	invoke(logger, level, t, map[string]interface{}{"key": "value"})
+	logger.Log("key", "value")
 
 	lastEntry := hook.LastEntry()
 
@@ -132,14 +69,13 @@ func testContext(level string, t *testing.T, args ...interface{}) {
 	}
 }
 
-func testMessageAndContext(level string, t *testing.T, args ...interface{}) {
+func TestLogger_MessageAndContext(t *testing.T) {
 	testLogger, hook := test.NewNullLogger()
-	testLogger.Level = _logrus.DebugLevel
 	logger := &logrus.Logger{
 		Logger: testLogger,
 	}
 
-	invoke(logger, level, t, "Test message", map[string]interface{}{"key": "value"})
+	logger.Log("msg", "Test message", "key", "value")
 
 	lastEntry := hook.LastEntry()
 
@@ -156,38 +92,30 @@ func testMessageAndContext(level string, t *testing.T, args ...interface{}) {
 	}
 }
 
-func testInvalid(level string, t *testing.T) {
+func TestLogger_Level_Debug(t *testing.T) {
+	testLevel(_logrus.DebugLevel, t)
+}
+
+func TestLogger_Level_Info(t *testing.T) {
+	testLevel(_logrus.InfoLevel, t)
+}
+
+func TestLogger_Level_Error(t *testing.T) {
+	testLevel(_logrus.ErrorLevel, t)
+}
+
+func testLevel(level _logrus.Level, t *testing.T) {
 	testLogger, hook := test.NewNullLogger()
 	testLogger.Level = _logrus.DebugLevel
 	logger := &logrus.Logger{
 		Logger: testLogger,
 	}
 
-	invoke(logger, level, t, "invalid 1", "invalid 2", "invalid 3")
+	logger.Log("level", level.String())
 
 	lastEntry := hook.LastEntry()
 
-	if lastEntry.Message != "" {
-		t.Error("message is expected to be empty")
-	}
-
-	if len(lastEntry.Data) != 0 {
-		t.Error("context is expected to be empty")
-	}
-}
-
-func invoke(logger *logrus.Logger, level string, t *testing.T, args ...interface{}) {
-	switch level {
-	case "Debug":
-		logger.Debug(args...)
-
-	case "Info":
-		logger.Info(args...)
-
-	case "Error":
-		logger.Error(args...)
-
-	default:
-		t.Fatalf("unknown logging level: %s", level)
+	if lastEntry.Level != level {
+		t.Errorf(`message is expected to have %s level, received: %s`, level.String(), lastEntry.Level.String())
 	}
 }